% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insol.R
\name{AnnualMean}
\alias{AnnualMean}
\title{Annual Mean insolation}
\usage{
AnnualMean(orbit, S0 = 1365, lats = seq(-90, 90, 1), degree = TRUE)
}
\arguments{
\item{orbit}{Output from a solution, such as \code{ber78}, \code{ber90} or 
\code{la04}}

\item{S0}{Total solar irradiance}

\item{lats}{list of latitudes at which annual mean is computed}

\item{degree}{true if latitudes are provided in degrees}
}
\value{
an object of class "AnnualMean" with annual mean insolations
}
\description{
Compute annual mean insolation for a given orbit and solar constant
}
