\name{BER78}
\alias{Table4}
\alias{Table1}
\alias{Table5}
\alias{BER90}
\alias{Table4_90}
\alias{Table1_90}
\alias{Table5_90}
\alias{.loaded}

\docType{data}
\title{Tables supplied by BER78 and BER90}
\description{
   Tables necessary for computation of astronomical solutions by
   Berger (1978) (in Table1, Table4 and Table5) and
   Beger and Loutre (1990) (in Table1_90, Table4_90 and Table5_90)
   }
\usage{
 data(BER78) 
 data(BER90) 
 }
\format{text files reproducing the published tables with original units.}

\references{
  Berger, A. L. (1978).  Long-term variations of daily
insolation and Quaternary climatic changes, J. Atmos. Sci., 35,
2362-2367.

  Berger and M.F. Loutre (1991), Insolation values for the climate of the last 10 million years,  Quaternary Science Reviews, 10, 297 - 317. }

  
\author{Michel Crucifix, U. catholique de Louvain, Belgium.}

\source{ftp://ftp.elic.ucl.ac.be/berger/berger78/ 
        ftp://ftp.elic.ucl.ac.be/loutre/QSR/ }

\examples{ 
data(BER78)
data(BER90)
}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{misc}
