% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BER90}
\alias{BER90}
\title{Tables supplied by BER90}
\format{
Text files reproducing the published tables with original units.
}
\source{
ftp://ftp.elic.ucl.ac.be/loutre/QSR/
}
\usage{
data(BER90)
}
\description{
Tables necessary for computation of astronomical solutions by
\insertCite{berger1990origine;textual}{palinsol}.
}
\note{
Table 2 is reconstructed based on the electronic files supplied
by the authors, based on the method  provided by
\insertCite{berger1990origine;textual}{palinsol} below.
}
\examples{
data(BER90)
}
\references{
\insertAllCited{}
}
\author{
Michel Crucifix, Université catholique de Louvain, Belgium.
}
\keyword{datasets}
