% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astro.R
\name{coprecession}
\alias{coprecession}
\title{Coprecession parameter, \eqn{e cos(varpi)}}
\usage{
coprecession(t, solution = ber78)
}
\arguments{
\item{t}{Time, years after 1950.}

\item{solution}{Solution used. One of \code{ber78}, \code{ber90} or
\code{la04}.}
}
\value{
Coprecession parameter, \eqn{e cos(varpi)}
}
\description{
Coprecession parameter, \eqn{e cos(varpi)}
}
\keyword{internal}
