% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{develop.discreteSpectrum}
\alias{develop.discreteSpectrum}
\title{Discrete spectrum reconstruction}
\usage{
\method{develop}{discreteSpectrum}(
  M,
  times = NULL,
  start = NULL,
  end = NULL,
  deltat = NULL,
  dfunction = cos,
  sum = TRUE
)
}
\arguments{
\item{M}{: discreteSpectrum object}

\item{deltat}{: see start.}

\item{dfunction}{is the trigonometrical function. Classically one of 'cos', 'sin', or 'cis'}

\item{sum}{: TRUE if user wants to sum components %in% the reconstruction}

\item{times:}{if supplied, times of the decomposition}

\item{start:}{if supplied, overrides time and will generate a time series with start and deltat, which must then
be supplied as well}
}
\value{
list of reconstructed components if sum=FALSE,  full
        reconstructed time series otherwise
}
\description{
Discrete spectrum reconstruction
}
\note{
if none if times, start and deltat are supplied, will reconstruct based on the attribute `xdata`
        which must then be present. If no `xdata` is availble, return an error.
}
