% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/develop.R
\name{develop}
\alias{develop}
\title{Develop a spectrum into a time series (generic)}
\usage{
develop(M, start = NULL, end = NULL, deltat = NULL, times = NULL, ...)
}
\arguments{
\item{deltat}{: see start.}

\item{arg:}{input class}

\item{times:}{if supplied, times of the decomposition}

\item{start:}{if supplied, overrides time and will generate a time series with start and deltat, which must then
be supplied as well}
}
\value{
nothing
}
\description{
Develop a spectrum into a time series (generic)
}
\note{
place holder for type-specific develop functions
}
