% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raw-data.R
\name{generate_table2}
\alias{generate_table2}
\title{Re-generates Table 2 from (Berger and Loutre, 1990)}
\usage{
generate_table2(Table1, Table4, Table5, sol = "BER78")
}
\arguments{
\item{Table1}{Tibble object.}

\item{Table4}{Tibble object.}

\item{Table5}{Tibble object.}

\item{sol}{Solution name, e.g., 'BER78'.}
}
\value{
Tibble with the precession amplitudes and frequencies.
}
\description{
Re-generates the precession amplitudes and frequencies based on files
provided by (Berger, 1978), (labelled Table 2 in the original publication).
The present routine is an implementation of the original article (Berger
and Loutre, 1990)
}
\references{
Berger, A.L., 1978. Long-term variations of daily insolation and Quaternary climatic changes. Journal of Atmospheric Sciences, 35(12), pp.2362-2367.

Berger, A.L. and Loutre, M.F., 1990. Origine des fréquences des éléments astronomiques intervenant dans le calcul de l'insolation. Bulletins de l'Académie Royale de Belgique, 1(1), pp.45-106.
}
\keyword{internal}
