% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LA04}
\alias{LA04}
\title{Astronomical elements supplied by Laskar et al. 2004}
\format{
Text files.
}
\source{
http://www.imcce.fr/Equipes/ASD/insola/earth/earth.html
}
\usage{
data(LA04)
}
\description{
Astronomical elements (longitude of perihelion, obliquity and eccentricity)
supplied by \insertCite{laskar2004long;textual}{palinsol} by step of 1ka,
from -51Ma to present (la04past) and from present to + 21Ma.
}
\examples{
data(LA04)

}
\references{
\insertAllCited{}
}
\author{
Michel Crucifix, Université catholique de Louvain, Belgium.
}
\keyword{datasets}
