% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astro.R
\name{obliquity}
\alias{obliquity}
\title{Obliquity}
\usage{
obliquity(t, solution = ber78, degree = FALSE)
}
\arguments{
\item{t}{Time, years after 1950.}

\item{solution}{Solution used. One of \code{ber78}, \code{ber90} or
\code{la04}.}

\item{degree}{Boolean flag to indicate whether or not to return angles in
degrees.}
}
\value{
Obliquity
}
\description{
Obliquity
}
\keyword{internal}
