% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palinsol-package.R
\docType{package}
\name{palinsol-package}
\alias{palinsol}
\alias{palinsol-package}
\title{palinsol: Insolation for Palaeoclimate Studies}
\description{
R package to compute Incoming Solar Radiation (insolation) for palaeoclimate studies. Features three solutions: Berger (1978), Berger and Loutre (1991) and Laskar et al. (2004). Computes daily-mean, season-averaged and annual means for all latitudes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bitbucket.org/mcrucifix/insol}
}

}
\author{
\strong{Maintainer}: Michel Crucifix \email{michel.crucifix@uclouvain.be}

}
\keyword{internal}
