% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Insol.R
\name{plot.Milankovitch}
\alias{plot.Milankovitch}
\title{plot Milankovitch graph}
\usage{
\method{plot}{Milankovitch}(
  x,
  months = TRUE,
  polar_night = TRUE,
  plot_function = contour,
  col = "black",
  ...
)
}
\arguments{
\item{x}{Milankovitch object}

\item{months}{if true : x-axis of the plot indicates months conventionnally
defined with the true solar longitude; x-axis is simply the true solar
longitude otherwise}

\item{polar_night}{if true : the polar night zone will be hashed}

\item{plot_function}{function used to plot the matrix. Typically
\code{contour} or \code{image} but may also be \code{image.plot} if using
the \code{fields} package.}

\item{col}{main color for positive contours (negative contours are in red)}

\item{...}{Other arguments passed to plotting function}
}
\description{
plot Milankovitch object
}
\author{
Michel Crucifix, U. catholique de Louvain, Belgium.
}
\keyword{misc}
