% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astro.R
\name{precession}
\alias{precession}
\title{Precession parameter, \eqn{e sin(varpi)}}
\usage{
precession(t, solution = ber78)
}
\arguments{
\item{t}{Time, years after 1950.}

\item{solution}{Solution used. One of \code{ber78}, \code{ber90} or
\code{la04}.}
}
\value{
Precession parameter, \eqn{e sin(varpi)}
}
\description{
Precession parameter, \eqn{e sin(varpi)}
}
\keyword{internal}
