\name{BER78}
\alias{Table4}
\alias{Table2}
\alias{Table1}
\alias{Table5}
\alias{BER90}
\alias{.loaded}

\docType{data}
\title{Tables supplied by BER78 and BER90}
\description{
   Tables necessary for computation of astronomical solutions by
   Berger (1978)  or Berger and Loutre (1990). 
   }
\usage{
 data(BER78) 
 data(BER90) 
 }
\format{text files reproducing the published tables with original units.}

\note{Table 2 is reconstructed based on the electronic files supplied 
by the authors, based on the method  provided by Berger and Loutre (1990)
below. }

\references{
  Berger, A. L. (1978).  Long-term variations of daily
insolation and Quaternary climatic changes, J. Atmos. Sci., 35,
2362-2367, doi:10.1175/1520-0469(1978)035<2362:LTVODI>2.0.CO;2

    A. Berger and M. F. Loutre (1990), Origine des fr\'equences des \'el\'ements astronomiques intervenant dans l'insolation,  Bull. Classe des Sciences, 1-3, 45-106 

  Berger and M.F. Loutre (1991), Insolation values for the climate of the last 10 million years,  Quaternary Science Reviews, 10, 297 - 317, doi:10.1016/0277-3791(91)90033-Q

}
\author{Michel Crucifix, U. catholique de Louvain, Belgium.}

\source{ftp://ftp.elic.ucl.ac.be/berger/berger78/ 
        ftp://ftp.elic.ucl.ac.be/loutre/QSR/ }

\examples{ 
data(BER78)
data(BER90)
}


% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{misc}
