% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eccentricity_spectrum.R
\name{eccentricity_spectrum}
\alias{eccentricity_spectrum}
\title{Trigonometrical development of eccentricity spectrum, based on an erbital solution}
\usage{
eccentricity_spectrum(G)
}
\arguments{
\item{epi}{: discrete spectrum of epi, as e.g. `La88$epi` based on `data(La88)`}

\item{inputfreq}{: artificially reduces the number of terms to take into account}
}
\value{
discrete spectrum of eccentricity
}
\description{
This is a fully transparent R-implementation of the trigonometric epansion of excentricity,
based on a trigonometrical epansio of (e,pi) (e.g., `data(La88)`).
order  by decreasing frequency
}
\note{
# this actually provides the (exact) spectrum of m(1+A/2-A^2/8+A/16) where A= e^2/m^2 - 1 , and m =  sqrt(sum(epi$Amp^2))
}
\references{
Berger A. and M. F. Loutre (1990), Origine des frequences des elements astronomiques intervenant dans le calcul de l'insolation, Bulletin de la Classe des sciences, (1) 45-106 doi:10.3406/barb.1990.38523
}
