% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-raw-data.R
\name{generate_table2}
\alias{generate_table2}
\title{Re-generates Table 2 from (Berger and Loutre, 1990)}
\usage{
generate_table2(Table1, Table4, Table5, sol = "BER78")
}
\arguments{
\item{Table1}{Tibble object (unused).}

\item{Table4}{Tibble object.}

\item{Table5}{Tibble object.}

\item{sol}{Solution name, e.g., 'BER78'.}
}
\value{
Tibble with the precession amplitudes and frequencies.
}
\description{
Re-generates the precession amplitudes and frequencies based on files
provided by \insertCite{berger1978long}{palinsol}, (labelled Table 2 in the original publication).
The present routine is an implementation of the original article \insertCite{berger1990origine}{palinsol}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
