% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_constants.R
\name{newcomb_parameters}
\alias{newcomb_parameters}
\title{Compute the Newcomb constant and its sensitivity to e^, according to the Sharav-Budnikova model featured by Berger 1973.}
\usage{
newcomb_parameters(
  CAC = CAC_ref,
  omega = omega_solar,
  c = c_ref,
  ece = ece_ref,
  earth_moon_distance = earth_moon_distance_ref
)
}
\arguments{
\item{CAC}{is (C-A)/C where A and C are Earth's moments of inertia aroud the plora and equatioral axis of inertia. Non-dimensional}

\item{omega}{is the rotational angular velocity of the Earth, measured in a geocentric framework, in radians per second}

\item{c}{is the inclination of the  lunar orbit on the ecliptic, in radians}

\item{ece}{: moon orbit excentricity}

\item{earth_moon_distance}{is expressed in km}
}
\value{
a list with "ell" and "P0", ready for use in `compute_tables`, in arc seconds per year.
}
\description{
--
}
\details{
Implementation of expressions for "ell" and "P0" p. 112 of Berger 1973. These variables
are fundamental in the computation of the luni-solar precession (also called axial precession)
In first approximation, the precession frequency (the term k) appearing in astronomical developments
is equal to "ell"*cos(e_bar), where e_bar is the reference value for obliquity. 
d(ell)/d(e^2) = 3/2 P0, where "e" is the obliquity. All parameters are given default values
coresponding to the present-day state.
}
\examples{
# we consider the AstroGeo model of Farhat et al., Astron. Asttroph. 665
# length of day at -370 Ma : 21.81
# semi-major axis of moon: 361000km
N = newcomb_parameters (omega = 2*pi/(21.81*3600), earth_moon_distance = 361000)
print (N)
# assuming, still according to the AstroGeo22 model, an obliquity of 22.2 degrees, 
# this simple model suggests a precession rate of
N$ala * cos(22.2 * pi/180)
# MORE WORK IS NEEDED HERE BECAUSE  C-A/C also varies with OMEGA

}
\references{
Berger, A. L. (1973)
Berger A. L. (1973), The astronomical theory of paleoclimates, Theses, UCLouvain, 1973. Currently not available in digital version. 
Berger A., M. Loutre and V. Dehant (1989), Astronomical frequencies for pre-Quaternary palaeoclimate studies, Terra Nova, (1) 474–479 doi:10.1111/j.1365-3121.1989.tb00413.x
}
\author{
Michel Crucifix, UC Louvain, Belgium
}
