% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io2mtoiom.R
\name{SB67}
\alias{SB67}
\title{Sharaf-Budnikova model, based on Berger's implementation}
\usage{
SB67(
  PlanetarySolution,
  P0 = 17.3919 * pi/(3600 * 180),
  ell = 54.9066 * pi/(3600 * 180),
  epsilonbar = 23.399935 * pi/180,
  zeta = 1.600753 * pi/180,
  aggregating = TRUE
)
}
\arguments{
\item{P0}{: "newcomb" constant}

\item{ell}{: partial derivative of `P0` with respect to eccentricity}

\item{epsilonbar}{: reference obliquity}

\item{zeta}{: precession phase at year zero (note toself: I need to be a bit more accurate about the meaning here"}

\item{aggregating}{: should different terms with same frequencies be aggregated ? (default : `TRUE`)}

\item{orbital}{solution (e.g., `data(La88)`)}
}
\value{
a list with `discreteSpectrum` for OBLIQUITY, PSI (longitude of perihelion on fixed plane) and CLIMPRECESS
        a `discreteSpectrum` is a list with `Amp`, `Freq` (angluar velocity) and `Phases`, with attributes given the
        `shift` with respect to zero, and `trend` (for PSI).
}
\description{
This is a fully transparent R-recording of André Berger's implementation of the Sharak and Budnikova precession
model. Using, as inputs, a trigonometric expansion of (e,pi) and (i/2, omega) (e.g. `data(La88)`)
newcomb canstants (as, e.g., supplied by `newcomp_parameters`), mean obliquity and mean precession phase at
the reference year corresponding to the orbital solution (usually 1950.0 or 2000.0). 
Supplies trigonometrical expansions for obliquity (OBLIQUITY), longitude of the vernal equinox with respect to the fixed
plane (PSI) and climatic precession.
}
\note{
Running this routine typically takes one second on a modern computer, and precession and obliquity may then
       be computed very efficiently for any period of time where the supplied parameters are considered as valid
       However, given the nature of orbital solutions and also the range of validity of the SB development (accurate
       up to order 2 in eccentricity), this generally would not span more than 2 or 3 million years. This would be 
       enough for the recent past, or for illustrative purposes in the deep past. Modern codes for precession, more
       accurate, would however not take much more computing time in practice. 
       if aggregating is "FALSE", then the group, as per the arithmetic expansion of Berger (1973) and Berger and Loutre (1991)
       belong, is given.
}
\examples{

data(La88)
S <- SB67(La88, aggregating = TRUE)
S2 <- SB67(La88, aggregating = FALSE)

psi1 <- develop(S$PSI, -1e6, 0, 1e3)
psi2 <- develop(S2$PSI, -1e6, 0, 1e3)
plot(psi1)
lines(psi1 - psi2)
}
\references{
Berger, A. L. (1973), Berger and Loutre (1991), Sharaf S. G. and N. A. Boudnikova (1967), Secular variations of elements of the Earth's orbit which influences the climates of the geological past, Tr. Inst. Theor. Astron. Leningrad, (11) 231-261
}
